Add a reaction listener, generalize class discovery, add a context
To get the necessary objects to the commands and watchers, which are
dynamically instantiated, they were combined in a `Context` object.

This object is passed around when the two are called.

In order to create this object the command listener and the reaction
listener are needed, so they get the context injected later on.
This is ugly but works.